agents/trading-system/requirements.txt
# Multi-Agent Trading System dependencies
ccxt>=4.3.0
langgraph>=0.2.0
langchain-core>=0.2.0
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
pyyaml>=6.0
